Safari-609.1.11.1/LayoutTests/fast/forms/listbox-visible-size-expected.txt

PASS select.scrollTop is targetTop
PASS successfullyParsed is true

TEST COMPLETE
